sacme/requirements.txt

django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
Pillow==7.1.2
requests==2.23.0
xlrd==1.2.0
mysqlclient==2.0.1
